Разбивает проект на компоненты
